Rearranges some files.
