+++ app/src/main/res/layout/content_main.xml
@@ -55,7 +55,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/main_txt_subtitle"
         app:layout_constraintVertical_bias="0.0"
-        tools:listitem="@layout/holder_deal" />
+        tools:listitem="@layout/item_deal" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
