+++ app/src/main/res/layout/item_deal.xml
@@ -63,6 +63,13 @@
             app:layout_constraintTop_toTopOf="@+id/deal_txt_discount"
             tools:text="@string/placeholder_value" />
 
+        <View
+            android:background="@color/colorAccent"
+            android:layout_width="match_parent"
+            android:layout_height="1dp"
+            app:layout_constraintStart_toStartOf="@id/deal_txt_value"
+            app:layout_constraintTop_toBottomOf="@id/deal_txt_value"/>
+
     </androidx.constraintlayout.widget.ConstraintLayout>
 
 </androidx.cardview.widget.CardView>
